Created search video component
