+++ src/components/Nav/Nav.styles.css
@@ -2,27 +2,22 @@
   display: flex;
   flex-direction: row;
   width: 100%;
-  justify-content: space-around;
+  justify-content: space-evenly;
 
   /* shadow */
-  -webkit-box-shadow: 0px 5px 10px 0px rgba(128,216,247,0.3);
-  -moz-box-shadow: 0px 5px 10px 0px rgba(128,216,247,0.3);
-  box-shadow: 0px 5px 10px 0px rgba(128,216,247,0.3);
+  -webkit-box-shadow: 0px 5px 10px 0px rgba(128, 216, 247, 0.3);
+  -moz-box-shadow: 0px 5px 10px 0px rgba(128, 216, 247, 0.3);
+  box-shadow: 0px 5px 10px 0px rgba(128, 216, 247, 0.3);
 }
 
 .topnav > * {
   margin: 20px;
 }
 
-button {
-  background: #000;
-  color: white;
-  border: 0px;
-  font-size: 1.125rem;
+p {
   font-weight: bold;
 }
 
 .topnav > *:hover {
-  color: #80D8F7;
-  font-size: 1.12501rem;
+  color: #80d8f7;
 }
