+++ src/components/SearchVideo/index.js
@@ -1,0 +1,1 @@
+export { default } from './SearchVideo.component';
